functions/conf : `XDG_CONFIG_HOME` manager
Open or edit configuration file under `XDG_CONFIG_HOME`, by default,
it is use `nvim` as editor and `nvim -c :Telescope find_files` as
directory opener.

The arguments is the config directory or file, if given more than one
arguments, then the arguments will be treated as subdirectory, if the
last arguments is a file, then it will opened using the `edit` variable,
if a directory, the `open_editor` variable will be used.

example :

```sh
$ conf fish conf.d # will open $XDG_CONFIG_HOME/fish/conf.d/

$ # will edit $XDG_CONFIG_HOME/fish/functions/conf.fish
$ conf fish functions conf.fish
```

On exit, it will throw back to the previous `$PWD`.