functions/dotfile : Initial implementation, dotfile manager
Use bare repository to manange my dotfiles, by default, it's use
`~/git/dotfiles` directory as `$GIT_DIR`, so to use this function as it
is, you must set your `$HOME` as bare repository with the following
command:

    ```
    cd ~/
    git init --bare $HOME/git/dotfiles
    dotfile config --local status.showUntrackedFiles no
    dotfile remote add <remote> <url>
    ```

If you want to change the `$GIT_DIR`, change the `DOTFILES` variable
inside the `__dotfile_cmd` function and use it as `--bare` value, or
if you already have a repository, go to that repository and run this
command: `git config --local core.bare true`.

Then, run the previous command to init your `$HOME` as bare repository.